Completed 203 v3
The better iterative approach was quite easy to understand. I knew beforehand that copying a variable would share the same reference as the original object, but I didn't know that the new variable can be reassigned without affecting the original. So in this situation, a new pointer can be used to check to see if the current node.val == val, but then reassigned to check the subsequent nodes.