Delay 110
I think that it would make more sense to do 111. Minimum Depth of Binary Tree first because the condition states that "a binary tree in which the left and right subtrees of every node differ in height by no more than 1," meaning that I should find the difference between the maximum and minimum depth of the tree.